Minor updates to state_controller.py.
Also added a new PID class to state_contoller.py. This was for a class
project, and 1) should definitely not live in state_controller.py
and 2) Is probably not any cleaner than the existing implementation.
However, the rest of state_controller is much better.

Terminal_mode_selector provides a way to test changing between states,
and it provides a simple way to send commanded position to the drone from
the commandline