Mainly improvements to 2D UKF.
Got rid of prediction in the 2D UKF's IMU callback, keeping it to the IR 
callback to only occur right before performing a measurement update. 
Added a little bit of "adaptive" filtering to the 2D UKF, as well as a 
check to ensure a non-negative time step. Also improved the 2D drone 
simulator to publish simulated camera pose data. Furthermore, moved the 
call to rospy.Time.now() in the IR and flight controller nodes.