Deleted throttle memory between flights
The roll and pitch init_i values are stored between flights so the low i-terms can converge faster. This was being done for the throttle as well, but this was causing the drone to take off too quickly after the first flight. The low_i term needed to take off the drone is less than what the value converges too while flying. Thus, we no longer store the throttle value so that the drone can take off at the same rate with each flight.